added pinging functions

+++ models/pings.js
@@ -2,7 +2,7 @@
 
 module.exports = function(sequelize, DataTypes) {
 	var PING = sequelize.define('PING', {
-		delay: { type: DataTypes.INTEGER, allowNull: false },
+		delay: { type: DataTypes.DECIMAL, allowNull: false },
 		state: { type: DataTypes.STRING, allowNull: false }
 	}, {
 		classMethods: {
